requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
deluge-client==1.10.2
idna==3.8
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
